tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/034.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: innerHTML adding frames with JS in</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>
  <div></div>
  <script>log('inline script #1');
    document.getElementsByTagName('div')[1].innerHTML = '<iframe src="pages/helloworld.html"></iframe>';
    log('end script #1');
  </script>

  <script type="text/javascript">
  log( 'inline script #2' );
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'inline script #2', 'frame/popup script']);
    t.done();
  }
  onload = t.step_func(function(){setTimeout(t.step_func(test), 200);})
  </script>

</body></html>
